requirements.txt
PyQt5
os-sys
google-generativeai